+++ Models/Enrollment.cs
@@ -2,9 +2,9 @@
 {
     public class Enrollment
     {
-        public long Id { get; set; }
-        public long StudentId { get; set; }
-        public long CourseId { get; set; }
+        public int Id { get; set; }
+        public int StudentId { get; set; }
+        public int CourseId { get; set; }
         public DateTime EnrolledAt { get; set; } = DateTime.UtcNow;
 
         // Navigation properties
